Fix reload problem
Fixed reload problem when you are in a details page.  Before, it is
possible that the coffeelist hadn't been loaded so when you are looking
at the details for specific coffee, you get nothing.  Now it routed to
the list if you get nothing and then comes back when you have list of
something.

Also added handling for the pageinit event since this is jquerymobile.
will be cleaning up that code later for better mvc.