Page Init is now modularized
Because we are using JQM to render the pages, we need to find an event
that allows up to manipulate the DOM just before the page is show.  So
we bind the PageInit event with function pageInitHandler which finds out
router and calls its pageInitHandler.  The router's pageInitHandler
looks for the currentPage which is set when you call slidePage to render
page.  Then currentPage's pageInitHandler is called which is defined for
all Backbone.View in the prototype.