Fixed reload issue for edit and details
a curious things happens when you reload in the edit or details page.
sometimes, the coffee list has not been fetched properly so the
requested id for the coffee doesn't show up.  so we have to call the
list function in the router and then the list function needs to called
the orginating function again.