Added autocomplete when you are entering roasters in the add page
When users are in the add page and they are entering a roaster, we try
to do an autocomplete from a list of roasters.  Right now the list is
hard coded, the plan is to start with a seeded list but as user adds new
roasters, they are added to the list.